improve reliability of clicking

+++ src/input.h
@@ -1,9 +1,9 @@
 #pragma once
 #include "../game/shared.h"
 
-bool KeyEvent(int key, bool down, unsigned time);
-bool MouseEvent(int button, bool down, unsigned time);
-bool JoyEvent(int controller, int button, bool down, unsigned time);
+bool KeyEvent(int key, bool down, double time);
+bool MouseEvent(int button, bool down, double time);
+bool JoyEvent(int controller, int button, bool down, double time);
 const char *IN_BindForKey(int key);
 void IN_KeyDown(kbutton_t *b);
 void IN_KeyUp(kbutton_t *b);
